requirements.txt
fastapi
uvicorn
psycopg2
pymongo
python-dotenv
google-generativeai
matplotlib
